TBenchmark is moved to separate file
